+++ libs/ui/listbox/README.md
@@ -1,0 +1,7 @@
+# listbox
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test listbox` to execute the unit tests.
