+++ libs/ui/listbox/src/lib/listbox-option/listbox-option.component.ts
@@ -3,107 +3,12 @@ import {
   Component,
   ElementRef,
   computed,
-  contentChildren,
-  effect,
   inject,
   input,
-  model,
-  output,
-  signal,
 } from '@angular/core';
 
-export enum Orientation {
-  Vertical,
-  Horizontal,
-}
+import { Listbox } from '../listbox/listbox.component';
 
-@Component({
-  selector: 'jm-listbox',
-  template: `<ng-content />`,
-  standalone: true,
-  host: {
-    role: 'listbox',
-    '[tabIndex]': 'disabled() || activeOption() ? -1 : 0',
-    '(focus)': 'activateOption()',
-    '[style.flex-direction]': 'orientationStyle()',
-    '(keydown)': 'navigate($event)',
-  },
-  styles: `
-    :host {
-      display: flex;
-      overflow-x: auto;
-      gap: 0.5rem;
-      padding: 1rem 2rem;
-      margin: 0 -2rem;
-    }
-    @media (max-width: 1000px) {
-      scrollbar-width: none;
-    }
-  `,
-})
-export class Listbox<T> {
-  readonly value = model<T[]>([]);
-  readonly disabled = input(false);
-  readonly orientation = input<Orientation>(Orientation.Horizontal);
-
-  readonly options = contentChildren(Option);
-  activeOption = signal<Option<T> | undefined>(undefined);
-
-  readonly orientationChange = output<Orientation>();
-
-  protected readonly orientationStyle = computed(() =>
-    this.orientation() === Orientation.Horizontal ? 'row' : 'column'
-  );
-
-  constructor() {
-    effect(() => {
-      this.orientationChange.emit(this.orientation());
-    });
-    effect(() => {
-      this.activeOption()?.focus();
-    });
-  }
-
-  navigate(event: KeyboardEvent) {
-    if (event.key.startsWith('Arrow')) {
-      event.stopPropagation();
-      event.preventDefault();
-    }
-    switch (event.key) {
-      case 'ArrowLeft':
-      case 'ArrowUp':
-        this.moveFocus(-1);
-        break;
-      case 'ArrowRight':
-      case 'ArrowDown':
-        this.moveFocus(1);
-        break;
-    }
-  }
-
-  moveFocus(moveBy: number) {
-    const activeOption = this.activeOption();
-
-    const nextPosition =
-      activeOption && moveBy
-        ? this.options().indexOf(activeOption) + moveBy
-        : 0;
-
-    if (nextPosition >= 0 && nextPosition < this.options().length) {
-      this.activeOption.set(this.options()[nextPosition]);
-    }
-  }
-
-  activateOption() {
-    const selectedOption = this.options().find((o) => o.isSelected());
-    if (selectedOption) {
-      this.activeOption.set(selectedOption);
-    } else {
-      const firstEnabledOption = this.options().find((o) => !o.disabled());
-      this.activeOption.set(firstEnabledOption);
-    }
-  }
-}
 
 @Component({
   selector: 'jm-option',
@@ -202,7 +107,7 @@ export class Option<T> {
     if (!this.isDisabled()) {
       const value = this.value();
 
-      this.listbox.value.update((lastOptions) => {
+      this.listbox.value.update((lastOptions: T[]) => {
         const newList = [...lastOptions];
         const optionsIndex = lastOptions.indexOf(value);
