+++ libs/ui/listbox/src/lib/models/orientation.ts
@@ -1,0 +1,4 @@
+export enum Orientation {
+  Vertical,
+  Horizontal,
+}
