--- src/app/presentation/modules/listbox/index.ts
@@ -1,2 +1,0 @@
-
-export * from './listbox';
